pull out parseKnownNetworks & add to utils
